fix fseek on x64

+++ src/cpio.cpp
@@ -129,7 +129,7 @@ int cpioGetFile(FILE* cpioArchiveFd, const char *file)
             // this is not the file we are looking for
             // skip the contents of this file
             if (filesize % 2 == 1) filesize++; // padding of 1 null byte
-            int r = fseek(cpioArchiveFd, filesize, SEEK_CUR);
+            int r = fseek(cpioArchiveFd, (long)filesize, SEEK_CUR);
             if (r != 0) {
                 LOG_ERROR("fseek error (file=%s contentsSize=%u): %s", filepath.c_str(), filesize, strerror(errno));
                 return -1;
@@ -168,7 +168,7 @@ int cpioExtractTree(FILE* cpioArchiveFd, const char *src, const char *dst)
         // if file does not match src, then skip this file
         if (src && strlen(src) && 0 != strncmp(src, filepath.c_str(), strlen(src))) {
             if (filesize % 2 == 1) filesize++; // padding of 1 null byte
-            int r = fseek(cpioArchiveFd, filesize, SEEK_CUR);
+            int r = fseek(cpioArchiveFd, (long)filesize, SEEK_CUR);
             if (r != 0) {
                 LOG_ERROR("fseek error (file=%s contentsSize=%u): %s", filepath.c_str(), filesize, strerror(errno));
                 return -1;
@@ -324,7 +324,7 @@ FILE *cpioOpenArchive(const char *file)
     LOG_DEBUG("cpioExtractFile: size=%u\n", size);
 
     // go to the beginning of the cpio archive
-    r = fseek(f, -4-size, SEEK_END);
+    r = fseek(f, (long)(-4-size), SEEK_END);
     if (r != 0) {
         LOG_ERROR("Cannot fseek %d file '%s': %s", -4-size, file, strerror(errno));
         return NULL;
